add data contributed by CRSP to the package
NOTE: this is an intermediate commit, this data will not be usable in the package until suitable binary rdata files and data set documentation are added to the package.

This commit contains data contributed under license to the FactorAnalytics package by the Center for Research in Security Prices at the University of Chicago's Booth School of Business (CRSP) and by S&P Global Markets.

The CRSP data is not covered by the GPL. Use of the data in derivative works is not permitted without the express written permission of CRSP.